Bug 509989 - Add some heuristics to find NodeJS and VSCode on macOS
For VSCode, it looks in /usr/share/code but on macOS, it is typically
installed in /Applications/Visual Studio Code.app. This patch changes
the default path to that on macOS.

To find NodeJS, it uses the output of "which node" to figure out
where the command is. By default, node is installed in
/usr/local/bin/node but it is only added to the PATH when running
from a terminal. When starting Eclipse from the UI (Dock, etc),
instead it uses the environment specified to launchctl which doesn't
contain by default /usr/local/bin. This patch adds a fallback to try
to use /usr/local/bin/node on macOS if it exists.

Change-Id: I9dad821de2d8cdfe5dbc936ee45809da2d730fea
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>