Bug 513322 - ProcessStreamConnectionProvider doesn't read from error stream causing server to block
When using ProcessStreamConnectionProvider to implement a language
server integration, the input streams and output streams are used but
the error streams is not. Because ProcessBuilder creates a pipe by
default, the stream has to be read otherwise the process might write to
the stream and block once the buffer is full. One option is to just set
the ProcessBuilder to inherit the error stream from the parent process
(Eclipse) and it will be properly read. One nice side effect is that you
see the error stream in the Console view when developing LSP4E.

Change-Id: I3e75c4388c4960fda2759b1192c746b3bc988335
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>