Bug 520055 - Marker resolution doesn't delegate to the server if it supports command execution
If code actions can be performed on an LSP diagnostic, it will have a
command id associated with it. When resolving the Eclipse IMarker via
a Quick Fix, we should check the command's id against the list of
commands that the language server states as being capable of
executing. If the language server can handle the IMarker's command,
delegate the resolution of the IMarker to the server via the
'workspace/executeCommand' request.

Change-Id: I910200eba3ce20c1d9d5f0548f1eff16a870b539
Signed-off-by: Remy Suen <remy.suen@gmail.com>