Bug 520771 allow to hide the 'kind' part in the outline view
The 'kind' property is currently limited to a few possibilities, which don't match the relevant concepts of a lot of languages.
So there are good reasons to not show it. Using the customization mechanism described in the JavaDoc of SymbolInformationPropertyTester, you can override the label provider. It is a good idea not to start from scratch then but to reuse a customized version of the original LabelProvider. This changes prepares for such reuse, as well as for an action to toggle the display of the kind property.

Signed-off-by: Jan Koehnlein <jan.koehnlein@typefox.io>
Change-Id: I62354105476132f103c96005f30216f6bb030f7a
Signed-off-by: Jan Koehnlein <jan.koehnlein@typefox.io>