Bug 520052 - Rename assumes that workspace edits are in reverse order
The text edits that are returned by a language server have no
ordering defined by the specification. The specification does however
state that when text edits are applied by the client, they must be
applied from the bottom to the top of the text document. To that end,
we should sort the text edits received from the client when turning
them into a CompositeChange so that they are applied to the document
correctly.

Change-Id: I176c4b493e484f1a93d5d5ed0d879f1a5bf49240
Signed-off-by: Remy Suen <remy.suen@gmail.com>