Fix two more breakage due to LSP4J api changes
These were missed in previous patch-set because
it is less obvious (no compilation error).

Firts one takes the toString of a 'Either' type
which isn't really something you want to
see in a completion's 'additional information'
control popup.

Second one is a similar implicit toString which
causes a test failure.

Change-Id: Ib5c586de639f774cb051ff5f9aa1aa3fdad221bc
Signed-off-by: Kris De Volder <kdevolder@pivotal.io>