Bug 508488 - APIs and some use-case for non IFiles
This introduce APIs that are on document level, working without IFile.
The document listener and Hover test are updated to support it.i

This also updates various internals of the wrapper or LanguageServiceAccessor.
While it would have been better in a separate change, those edits have been
helpful in development of this feature.

Change-Id: I0d7b92ec39b09641b2fd20e2c79a77a873ea1e81
Signed-off-by: Mickael Istria <mistria@redhat.com>