Use more then...Async future
This might reduce the wait duration. LSP4J locks on the current thread,
of if the thread/future that owns the request is used for further usage
(as part of a then...), it's preventing other requests to be processed.
Using then...Async should free the thread after response is retrieved,
making LSP4J request/response bus more available for other threads.

+ Tentative to make some tests more stable.

Change-Id: I3b001dac6b6d3f8407bcde69c238f792d2fbf43c
Signed-off-by: Mickael Istria <mistria@redhat.com>