Remove timeout on computeCompletionProposals
The timeout is very short and seems contrary to the
fact that the caller
org.eclipse.jface.text.contentassist.AsyncCompletionProposalPopup)
handles calling the proposal computers asynchronously.
AsyncCompletionProposalPopup shows a kind of 'progress / busy'
message, and updates this with real completions as soon as they
arrive.

Adding the timeout doesn't seem to provide any
additional value here. On the contrary, the timeout (of 1s)
is too short to even read the 'busy' message, and will replace
it with a nasty error message, even when the completion provider
may only need just a little bit more time.

Change-Id: I2dba516c96bfdb66b1dfde804cceb0b6ffbeab44
Signed-off-by: Kris De Volder <kdevolder@pivotal.io>