Bug 550478: Install toggle breakpoint handler less aggressively
We had been installing the toggle breakpoint handler (the thing that
makes clicking on the left hand column install a breakpoint) by registering
an adapter factory which could get a toggle breakpoint from any editor.

This worked, but was overly aggressive in that this toggle breakpoint
would become the default handler. That is, unless the editor provided
its own adapter to IToggleBreakpointsTarget. In the latter case, it would
not be possible to use LSP4E breakpoints at all.

This change contributes the toggle breakpoint via the org.eclipse.debug.ui.toggleBreakpointsTargetFactories
extension point instead. That allows all contributions to be collected and
presented in the "Breakpoint Types" context menu (or via Run) menu and does
not interfere. The two additional advantages of this are:
1- When the editor provides IToggleBreakpointsTarget adapter already, we
can now show the LSP4E version.
2- Rather than appearing in the UI as "Default" we can display it with our
own name.

Change-Id: I890ee57f3c10b997441daa459cc894f42ab96a72
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>