Bug 564070: Handle exception on launch/attach command
Previously a launch or attach would cause an exception, but
the thenCombineAsync that is next would be waiting for the
initialized to also be ready, however that was never going
to happen because of the exception on the launch/attach.

Therefore, if the launch/attach finishes exceptionally, mark
initialized as finishing exceptionally so that the sequence
does not hang.

Change-Id: I3c23fb4fd970475b94d8444b412de355491282dd