Allow DSPDebugTarget to complete earlier for non-debug launch
The Debug Adapter may not send the initialized notification in non-debug
mode but still work (for example the node debug adapater from Wild Web
Developer just run without further notifications if noDebug=true). So we
adapt the initialization process to that in case we're not debugging, we
avoid waiting for useless "initialized" notification and avoid sending
irrelevant debug-specific configuration such as breakpoint.

Change-Id: I87b081af6f3c3f71c875a0d71377866db9e26de9
Signed-off-by: Mickael Istria <mistria@redhat.com>