Bug 576425 - Support locations outside a local file system
Support locations outside a local file system by calling
IResource#getLocationUri() if IResource#getLocation() returns null, that
is the case for resources which are not in a local file system.

Since a File cannot be created from remote URIs, getDocumentContentTypes
is also adapted to retrieve an IFile compatible object directly from the
document instead of creating a new one from the URI just to get the name
of it.

Signed-off-by: Rubén Porras Campo <pcr@avaloq.com>
Change-Id: I46adce1390a86186133e74406dce59451f05dd89