Bug 576425 - Support finding resources for remote URIs
For remote URIs, where we do not have local files, we delegate
to an adapter to transform URIs to an IResource or an IFile.

Another option would be to iterate through all projects and
ask if they have a member for the given URI, but this may
not succeed without a URI mapper for most remote URIs, so
we delegate to an adapter which can do the URI mapping and
return the corresponding IResource or IFile.

Change-Id: I0362aecc8d5b905b481fa5a4f27d2c5f4f862112