Bug 576425 - Support finding resources for remote URIs
For remote URIs, where we do not have local files, we delegate to an
adapter to transform URIs to an IResource or an IFile, also when opening
the uri in a workbench page.

To do this, we call LSPEclipseUtils.findResourceFor(uri) even if we do
not have a file uri, and let if find the resource. We achieve this by
removing the constraint on the method open for the uri, and let the
method openFileLocationInEditor handle all what is not an HTTP uri.

In addition, test only for the INTRO URL if it is an HTTP URL. This
reduces the amount of work to open other URLs which are not an HTTP URL.