Choose 'best' resource when file url is ambiguous
In Eclipse workspace it is possible that a single file corresponds
to multiple workspace resources. This typically arises in 'multi-project'
setups where both the parent project and its nested children are
imported into the workspace.

This creates all sorts of confusions and subtle bugs.

The logical solution seems to be to deliberatly choose the
'most nested' resource instead of just arbitarily picking the
first one found.

This small does just that in LSPEclipseUtils.

This indirectly fixes issues in STS such as:

https://github.com/spring-projects/sts4/issues/769
https://github.com/spring-projects/sts4/issues/747