Fix failing tests
These tests grab a bitmap of a screen area and look for pixels of a particular colour, which is unreliable in the presence of things like anti-aliasing.
Fix by using a euclidian distance and tolerance for comparing colours, and by using the unicode solid block character to minimise the sorts of complicated edge that will get anti-aliased.