Concurrency: prevent race conditions when checking the doc version
Because we were checking the modification stamp inside the body of a
future, this could effectively happen on a random thread. This means
there is a chance that between the end of the future's computation and
the start of the code running on the Display thread that actually
updates the text, the user might make an edit.

As part of this I have audited the code and we should now only be
accessing the document's modification stamp from the SWT thread.