Improve marker resolution performance
by not connecting the file where we retrieve marker information. That is
an unnecessary step which can be expensive (for example the Xtext
language server rebuilds the source) and causes timeouts on
checkMarkerResoultion.

The side-effect of connecting the document when getting an initialized
server has been removed from all the
LanguageServerAccessor#getInitializedLangugeServer* methods.

Downstream project which rely on this behaviour should connect the
document after getting the initialized server.