formatting: Ensure that formatting not applied if the document has changed.
Implemented a general mechanism for wrapping results returned from the server which decorates
it with the version of the document at the time the call was dispatched. If it has been overtaken by subsequent
modifications then the caller can choose not to apply any changes suggested by the server. Effectively an
optimistic locking strategy but one in which it is up to the caller to decide whether to proceed or not.