Document Synchronization: Restrict internal hook
We want interaction with the language server for a given document to be sequenced correctly with document updates, so don't really
want to expose lastChangeFuture directly, even via a package-protected method. Since this method was used only by the LanguageServerWrapper
post-construct so that it could be used to ensure didOpen completed, have made this more explicit.