Improve logging of java.util.concurrent.TimeoutException
If the server is not responsive enough logging errors without any other
context than a stacktrace looks to the end user a bit scary, as if the
IDE would be broken.

Instead reporting a warning with a message that explains the context and
include the timeout that was not met.