Fix lost error message checkInitialConditions
The method checkInitialConditions was handling the error using an status
which was never returned. In addition the method handleError was
returning a value which was never used.

Both problem has been fixed by removing handleError and adding a method
that just gets the error message. This makes more clear where this error
message is attached. The error handling of checkInitialConditions has
been then fixed accordingly.