Inlay Hints: Use whole document rather than an empty range for the request
The initial implementation used an empty range at the start of the document. Language servers that respect the viewport range
will return no response. Language servers that ignore the range will return inlay hints for the whole document anyway.
Hence might as well set the range to the full text of the document, especially as eclipse only requests code minings when the
content is loaded or changed, not when the viewport changes.