Initial implementation of a semantic tokens reconciler
Initial implementation of a semantic tokens reconciler which supports
only SemanticTokensFull.

The reconciler uses the theme defined for TM4E so that the same styles
are used for the same token types.

In addition to the tokens defined for TM4E, also the deprecated modifier
is supported by marking striking out the regions.