Handle case where capabilities is still null during configurationDone
Debug adapters are not supposed to send initialized event until after
replying to initializeRequest with Capabilities. However some debug adapters
don't respect this (or didn't in the past). Even if they do respect this,
sometimes due to the multithreaded event handling, the initialized event
arrives before the capabilities are stored. Therefore use a future to guard
reading the capabilities before they are ready.

Fixes #51