Remove unnecessary - and incorrect - double asynchronous dispatch
Using chained futures with multiple async compositions isn't quite correct, even if they are all scheduled onto the same dispatcher.
If the connection future is completed then we need the language server operation we wish to execute to be scheduled straight onto
the dispatch queue. What we had was scheduling something onto the dispatch queue, that in turn scheduled the operation onto the
back of the dispatch queue, allowing another call (e.g. a document update) to interleave.