Prevent LSPFormatHandler.execute from trashing the editor contents
If between a format request and the moment the text edits are applied
the document has been edited, the text edit is discarded instead of
applied. Thus we prevent user edits which happened concurrent to the
request to be trashed without needing to lock the document or the UI.

When this happens the user gets informed with a notification.