Remove more then*Async
It seems like LS lookup or document state operation really need to
happen in a controlled flow/sequence and that we shouldn't let them
being scheduled more "randomly" with then*Async but instead ensure they
are chained with plain then* operations.
The *Async opertaions are more important once the LS-side work is
completed and transferring it to client-side processing (so the thread
is not blocked by Client-side consumption of the result).