Draft API: Port a couple of classes to use the LSPExecutors API
These examples were already changed into a very similar form by the earlier proposal in PR#251.

However, DocumentColorProvider/ColorInformationMining also exhibit a common but potentially undesirable pattern in the codebase - using the
response from the server to construct objects which embed the raw LanguageServer as fields - this is so that they can make subsequent calls
on the same server without going via LanguageServiceAccessor.

I have proposed using LanguageServerWrapper in this case so that subsequent calls can use the high-level API and not potentially subvert the threading/ordering guarantees.
This commit shows that this approach seems to work with trivial changes.

TODO: Decide what API LanguageServerWrapper should expose and whether it should be hidden behind an interface (or in a new, higher-level object to wrap the wrapper!)