Implement anyMatching as a non blocking call using a default timeout
If the default timeout is not met, we return that a server is matching.
Usually the pattern is to ask anyMatching before enabling a
functionality on a given document, but if the functionality is enabled
and there will always be follow-up request to use the functionality.

Usually the call is done to show/hide some menu item on the UI. It is
better to return quick and let the action do nothing if the user
actually runs the action than to block for a (potentially) long time the
creation on the menu items.