Handle empty quickfix list
Quick fixes for a problem marker shown fine if there are quick fixes
available. However, if there are no quick fixes available an empty list
of code actions is received on the lsp4e side. However, this result is
ignored (this the chunk of code is changed by this PR) and therefore
Computing... is shown indefnitely under quick fixes for problem that
don't have a quick fix.