Maintenance: Tidy up server lifecycle - fix workspace folder registration
Hadn't quite understood the existing logic. We cater for the fact that workspace folder capability can be dynamically registered
and unregistered but if the server responds that it supports them in the initial server capabilities on startup then we deem
that to be a permanent feature.

This is quite important to the server lifecycle as a wrapper gets kept even when the server it manages has shut down (e.g. from idle timeout).
Whether a server supports workspace folders is a key part of deciding whether a wrapper instance can be used for a particular document if
that document is from a different project than the one the wrapper was started with.