Final removal of all APIs that expose the language server object.
The test LanguageServiceAccessorTest.testStatusHandlerLSAsRunConfiguration
has been deleted as it cannot be reliably tested, it depends on creation
of the stream provider in LanguageServerWrapper#start, which happens in
a background thread.

The method LanguageServerWrapper#getVersion has been made public as a
replacement of the deleted class LSPDocumentInfo, which could be used to
indirectly get the version on a wrapper.