Add basic type hierarchy view
This is a basic implementation of a LSP based type hierarchy view with
the same info as in the exiting type hierarchy dialog which can be
opened by F4. In the next step the view could be expanded with filters
and a member info box.

Default Eclipse keys: F4 for Type Hierarchy and M1+T for Quick Type
Hierarchy

fixes #732