Make default implementation of configuration conform to specification (#792)
LSP Specification says "The order of the returned configuration settings
correspond to the order of the passed ConfigurationItems (e.g. the first
item in the response is the result for the first configuration item in
the params). [...] If the client can’t provide a configuration setting
for a given scope then null needs to be present in the returned array."