Ensure that active highlights are removed when a source view is closed
This commit resolves a bug related to erroneously persisted highlights
in source files. It includes a unit test that verifies the correctness
of the fix, and a few minor improvements to the testing framework.

---

Bug Description: Before this change, it was possible for highlights to
persist and be shown erroneously after a source view is closed. This
would happen, for example, after the following sequence of steps:

1. Open a source. Ensure that some keyword is highlighted.
2. Activate the option "Toggle Split Editor (Horizontal)".
3. In the newly opened source view, highlight a different keyword.
4. At this point, both sets of highlights (the one from step 1 and the
one from step 3) should be visible in both open source views. This is
correct behavior.
5. Close one of the source views, while keeping the other open. Before
this commit, both sets of highlights would persist. However, only the
set of highlights related to the source view that remains open should be
displayed.

Cause: Each instance of HighlightReconcilingStrategy adds/removes
highlights independently, and keeps track of all the highlights it has
created internally. However, all the highlights associated with the
source file's underlying IFileBuffer are only cleared when the file is
closed. If two source views display the same file, and one of them is
closed, its highlights should no longer be displayed.

Solution: When a source view is closed, all active highlights associated
with its HighlightReconcilingStrategy should be removed. We accomplish
this by calling `removeOccurrenceAnnotations()` during the `uninstall()`
method of `HighlightReconcilingStrategy.java`.

Testing: A unit test was created which approximately follows the
sequence of steps outlined above. To implement this unit test, the
following changes were made to the testing framework:

1. In HighlightTest.java, the function `assertAnnotationDoesNotExist`
was created (for consistency, the existing function
`assertHasAnnotation` was renamed to `assertAnnotationExists`).
2. The `mockDocumentHighlights` returned by the
`MockTextDocumentService` are now Position-dependent. The unit tests in
`HighlightTest.java` were adapted accordingly.