Semantic Tokens: Do not use default theme (#1029)
Fix for https://github.com/eclipse/lsp4e/issues/1028 - we requested the default theme if no TM mapping for the editor, but
this calls syncExec. Reconciler threads mustn't block on the UI thread as it is possible for the UI thread to be blocked
waiting for reconciliation; this deadlock was indeed observed.