Expand root outline view in outline view by default
Before this commit, the outline view is drawn with all elements
collapsed by default. This often results in plenty of wasted space in
the UI and low information density.

Instead, as discussed in #693, it would be preferable to draw the
outline view with its first level expanded by default (i.e., displaying
the root node and its immediate children by default).

Issue: #693