Fix disappearing sort button in outline view (fix #1091) (#1092)
Previously, the sort button was only active if the active part was the outline view (activePart had to be instanceof OutlineView). As soon as an editor became active, the visibleWhen expression became false and the sort button disappeared. The check was done by the expression definition named activePartHasCNFOutlinePage.

I added a new expression definition named activeEditorHasCNFOutlinePage. I changed the visibleWhen expression to only check if the (last) active editor (that does not necessarily has to be in focus) has an outline page of type CNFOutlinePage. The remaining checks using the expression activePartHasCNFOutlinePage are left unchanged.