Allow semantic highlighting to unset bold and italic
Semantic highlighting needs to be able to override the styles set by
TM4E. Simply not setting the styles in TM4E would introduce flickering
in the delay waiting for semantic highlighting.

If unsetting of bold and/or italic is not desired, disable using
semanticHighlightReconciler.ignoreBoldNormal and
semanticHighlightReconciler.ignoreItalicNormal

----

When applying semantic highlighting, use
TextPresentation#mergeStyleRanges to create new style ranges and merge
as "usual". Then traverse resulting style ranges and unset bold/italic
where it is not set in semantic highlighting.

The double traversal algorithm is simplified because after calling
TextPresentation#mergeStyleRanges we are guaranteed that each style
range from semantic highlighting has exact overlap with 1 or more style
ranges in the textPresentation.

----

Because textmate uses css, style {font-weight:"normal"} should be
expected to modify the font-weight, whereas {font-weight:"inherit"}
should not. However, TM4E parses the style into a an SWT StyleRange
which prevents the differentiation between "normal"=unset bold if it is
set and "inherit"=don't unset bold if it is set.

A more long term solution would use the css to determine whether an
unset bold bit is overriding or not.

For the medium term we simply say "semantic highlighting has ownership
of bold/italic and the style sheet must set them explicitly"